renamed donate to organizations, changed the route, renamed the component and files
