+++ src/App.tsx
@@ -14,8 +14,8 @@ import Register from "./components/accounts/Register";
 import SLogin from "./components/accounts/SLogin";
 import Profile from "./components/accounts/profile/Profile";
 // import ResetPassword from "./components/accounts/settings/resetPassword";
-import Donate from "./components/donate/Donate";
 import Home from "./components/home/Home";
+import Organizations from "./components/organizations/Organizations";
 import Favorites from "./components/pets/Favorites";
 import PetType from "./components/pets/PetType";
 import Pets from "./components/pets/Pets";
@@ -39,7 +39,7 @@ export default function App() {
               <Route path="about" element={<About />} />
               <Route path="resources" element={<Resources />} />
               <Route path="tips" element={<Tips />} />
-              <Route path="donate" element={<Donate />} />
+              <Route path="organization" element={<Organizations />} />
               {/* <Route path="stories" element={<Stories />} /> */}
               <Route path="favorites" element={<Favorites />} />
               <Route path="register" element={<Register />} />
